+++ HISTORY.rst
@@ -7,3 +7,11 @@ History
 ++++++++++++++++++
 
 * First release on PyPI.
+
+1.0 (25-08-2025)
+++++++++++++++++++
+
+* Stable release
+* Tested with Python 3.13
+* Support 'scipy.sparse' arrays.
+* Return 'mode="global"' coefficients as series instead of data frames with one row.